+++ desc/perturbations.py
@@ -8,8 +8,10 @@ from desc.backend import jnp, put, use_jax
 from desc.compute import profile_names
 from desc.objectives import (
     AxisRSelfConsistency,
+    AxisWSelfConsistency,
     AxisZSelfConsistency,
     BoundaryRSelfConsistency,
+    BoundaryWSelfConsistency,
     BoundaryZSelfConsistency,
     ObjectiveFunction,
     get_fixed_boundary_constraints,
@@ -55,11 +57,21 @@ def get_deltas(things1, things2):  # noqa: C901
             s1 = s1.copy()
             s2 = s2.copy()
             warnif(s1.M > s2.M or s1.N > s2.N, msg="The target surface" + msg)
-            s1.change_resolution(s2.L, s2.M, s2.N)
+            if hasattr(s1, "Mz") and hasattr(s2, "Mz"):
+                s1.change_resolution(s2.L, s2.M, s2.N, Mz=s2.Mz, Nz=s2.Nz)
+            else:
+                s1.change_resolution(s2.L, s2.M, s2.N)
             if not jnp.allclose(s2.R_lmn, s1.R_lmn):
                 deltas["Rb_lmn"] = s2.R_lmn - s1.R_lmn
             if not jnp.allclose(s2.Z_lmn, s1.Z_lmn):
                 deltas["Zb_lmn"] = s2.Z_lmn - s1.Z_lmn
+            if (
+                hasattr(s1, "W_lmn")
+                and hasattr(s2, "W_lmn")
+                and s1.W_lmn.size == s2.W_lmn.size
+                and not jnp.allclose(s2.W_lmn, s1.W_lmn)
+            ):
+                deltas["Wb_lmn"] = s2.W_lmn - s1.W_lmn
 
     if "axis" in things1:
         a1 = things1.pop("axis")
@@ -73,6 +85,8 @@ def get_deltas(things1, things2):  # noqa: C901
                 deltas["Ra_n"] = a2.R_n - a1.R_n
             if not jnp.allclose(a2.Z_n, a1.Z_n):
                 deltas["Za_n"] = a2.Z_n - a1.Z_n
+            if a1.W_n.size == a2.W_n.size and not jnp.allclose(a2.W_n, a1.W_n):
+                deltas["Wa_n"] = a2.W_n - a1.W_n
 
     for key, val in profile_names.items():
         if key in things1:
@@ -267,11 +281,23 @@ def perturb(  # noqa: C901
         Ainv = jnp.linalg.pinv(A)
         dc = deltas["Za_n"]
         tangents += jnp.eye(eq.dim_x)[:, eq.x_idx["Z_lmn"]] @ Ainv @ dc
+    if "Wb_lmn" in deltas.keys():
+        con = get_instance(constraints, BoundaryWSelfConsistency)
+        A = con.jac_unscaled(xz)[0]["W_lmn"]
+        Ainv = jnp.linalg.pinv(A)
+        dc = deltas["Wb_lmn"]
+        tangents += jnp.eye(eq.dim_x)[:, eq.x_idx["W_lmn"]] @ Ainv @ dc
+    if "Wa_n" in deltas.keys():
+        con = get_instance(constraints, AxisWSelfConsistency)
+        A = con.jac_unscaled(xz)[0]["W_lmn"]
+        Ainv = jnp.linalg.pinv(A)
+        dc = deltas["Wa_n"]
+        tangents += jnp.eye(eq.dim_x)[:, eq.x_idx["W_lmn"]] @ Ainv @ dc
     # all other perturbations besides the boundary
     other_args = [
         arg
         for arg in eq.optimizable_params
-        if arg not in ["Ra_n", "Za_n", "Rb_lmn", "Zb_lmn"]
+        if arg not in ["Ra_n", "Za_n", "Wa_n", "Rb_lmn", "Zb_lmn", "Wb_lmn"]
     ]
     if len([arg for arg in other_args if arg in deltas.keys()]):
         dc = jnp.concatenate(
@@ -469,6 +495,8 @@ def optimal_perturb(  # noqa: C901
     dPsi=False,
     dRb=False,
     dZb=False,
+    dW=False,
+    dWb=False,
     subspace=None,
     order=2,
     tr_ratio=[0.1, 0.25],
@@ -486,9 +514,10 @@ def optimal_perturb(  # noqa: C901
         Objective function to satisfy.
     objective_g : ObjectiveFunction
         Objective function to optimize.
-    dR, dZ, dL, dp, di, dPsi, dRb, dZb : ndarray or bool, optional
+    dR, dZ, dL, dp, di, dPsi, dRb, dZb, dW, dWb : ndarray or bool, optional
         Array of indices of modes to include in the perturbations of R, Z, lambda,
-        pressure, rotational transform, total magnetic flux, R_boundary, and Z_boundary.
+        pressure, rotational transform, total magnetic flux, R_boundary, Z_boundary,
+        omega, and omega_boundary.
         Setting to True (False) includes (excludes) all modes.
     subspace : ndarray, optional
         Transform matrix to give a subspace from the full parameter space.
@@ -541,11 +570,13 @@ def optimal_perturb(  # noqa: C901
         "R_lmn": dR,
         "Z_lmn": dZ,
         "L_lmn": dL,
+        "W_lmn": dW,
         "p_l": dp,
         "i_l": di,
         "Psi": dPsi,
         "Rb_lmn": dRb,
         "Zb_lmn": dZb,
+        "Wb_lmn": dWb,
     }
     deltas = {}
     for name, arg in argmap.items():
@@ -613,7 +644,7 @@ def optimal_perturb(  # noqa: C901
     xz = objective_f.unpack_state(jnp.zeros_like(xf), False)[0]
 
     for arg in deltas:
-        if arg not in ["Rb_lmn", "Zb_lmn"]:
+        if arg not in ["Rb_lmn", "Zb_lmn", "Wb_lmn"]:
             x_idx = eq.x_idx[arg]
             dxdc.append(jnp.eye(eq.dim_x)[:, x_idx])
         if arg == "Rb_lmn":
@@ -628,6 +659,12 @@ def optimal_perturb(  # noqa: C901
             Ainv = jnp.linalg.pinv(A)
             dxdZb = jnp.eye(eq.dim_x)[:, eq.x_idx["Z_lmn"]] @ Ainv
             dxdc.append(dxdZb)
+        if arg == "Wb_lmn":
+            con = get_instance(constraints, BoundaryWSelfConsistency)
+            A = con.jac_unscaled(xz)[0]["W_lmn"]
+            Ainv = jnp.linalg.pinv(A)
+            dxdWb = jnp.eye(eq.dim_x)[:, eq.x_idx["W_lmn"]] @ Ainv
+            dxdc.append(dxdWb)
     dxdc = jnp.hstack(dxdc)
 
     # 1st order